spliting test cases by scenario
