+++ src/NServiceBus.SqlServer.CompatibilityTests/Roundtrip.cs
@@ -9,118 +9,129 @@ namespace NServiceBus.SqlServer.CompatibilityTests
     using NUnit.Framework;
 
     [TestFixture]
-    public partial class MessageExchangePatterns
+    public partial class Roundtrip
     {
+        EndpointDefinition sourceEndpointDefinition;
+        EndpointDefinition destinationEndpointDefinition;
+
+        [SetUp]
+        public void SetUp()
+        {
+            sourceEndpointDefinition = new EndpointDefinition("Source");
+            destinationEndpointDefinition = new EndpointDefinition("Destination");
+        }
+
         [Test]
-        public void Send_1_2_to_2_2()
+        public void Roundtrip_1_2_to_2_2()
         {
             Action<IEndpointConfigurationV1> sourceConfig = c =>
             {
                 c.UseConnectionString(ConnectionStrings.Default);
-                c.MapMessageToEndpoint(typeof(TestCommand), "Destination");
+                c.MapMessageToEndpoint(typeof(TestRequest), "Destination");
             };
             Action<IEndpointConfigurationV2> destinationConfig = c =>
             {
                 c.UseConnectionString(ConnectionStrings.Default);
             };
 
-            VerifySend("1.2", sourceConfig, "2.2", destinationConfig);
+            VerifyRoundtrip("1.2", sourceConfig, "2.2", destinationConfig);
         }
 
         [Test]
-        public void Send_1_2_to_3_0()
+        public void Roundtrip_1_2_to_3_0()
         {
             Action<IEndpointConfigurationV1> sourceConfig = c =>
             {
                 c.UseConnectionString(ConnectionStrings.Default);
-                c.MapMessageToEndpoint(typeof(TestCommand), "Destination");
+                c.MapMessageToEndpoint(typeof(TestRequest), "Destination");
             };
             Action<IEndpointConfigurationV3> destinationConfig = c =>
             {
                 c.UseConnectionString(ConnectionStrings.Default);
             };
 
-            VerifySend("1.2", sourceConfig, "3.0", destinationConfig);
+            VerifyRoundtrip("1.2", sourceConfig, "3.0", destinationConfig);
         }
 
         [Test]
-        public void Send_2_2_to_1_2()
+        public void Roundtrip_2_2_to_1_2()
         {
             Action<IEndpointConfigurationV2> sourceConfig = c =>
             {
                 c.UseConnectionString(ConnectionStrings.Default);
-                c.MapMessageToEndpoint(typeof(TestCommand), $"Destination.{Environment.MachineName}");
+                c.MapMessageToEndpoint(typeof(TestRequest), $"Destination.{Environment.MachineName}");
             };
             Action<IEndpointConfigurationV1> destinationConfig = c =>
             {
                 c.UseConnectionString(ConnectionStrings.Default);
             };
 
-            VerifySend("2.2", sourceConfig, "1.2", destinationConfig);
+            VerifyRoundtrip("2.2", sourceConfig, "1.2", destinationConfig);
         }
 
         [Test]
-        public void Send_2_2_to_3_0()
+        public void Roundtrip_2_2_to_3_0()
         {
             Action<IEndpointConfigurationV2> sourceConfig = c =>
             {
+                c.MapMessageToEndpoint(typeof(TestRequest), "Destination");
                 c.UseConnectionString(ConnectionStrings.Default);
-                c.MapMessageToEndpoint(typeof(TestCommand), "Destination");
             };
             Action<IEndpointConfigurationV3> destinationConfig = c =>
             {
                 c.UseConnectionString(ConnectionStrings.Default);
             };
 
-            VerifySend("2.2", sourceConfig, "3.0", destinationConfig);
+            VerifyRoundtrip("2.2", sourceConfig, "3.0", destinationConfig);
         }
 
         [Test]
-        public void Send_3_0_to_1_2()
+        public void Roundtrip_3_0_to_1_2()
         {
             Action<IEndpointConfigurationV3> sourceConfig = c =>
             {
                 c.UseConnectionString(ConnectionStrings.Default);
-                c.RouteToEndpoint(typeof(TestCommand), $"Destination.{Environment.MachineName}");
+                c.RouteToEndpoint(typeof(TestRequest), $"Destination.{Environment.MachineName}");
             };
             Action<IEndpointConfigurationV1> destinationConfig = c =>
             {
                 c.UseConnectionString(ConnectionStrings.Default);
             };
 
-            VerifySend("3.0", sourceConfig, "1.2", destinationConfig);
+            VerifyRoundtrip("3.0", sourceConfig, "1.2", destinationConfig);
         }
 
         [Test]
-        public void Send_3_0_to_2_2()
+        public void Roundtrip_3_0_to_2_2()
         {
             Action<IEndpointConfigurationV3> sourceConfig = c =>
             {
                 c.UseConnectionString(ConnectionStrings.Default);
-                c.RouteToEndpoint(typeof(TestCommand), "Destination");
+                c.RouteToEndpoint(typeof(TestRequest), "Destination");
             };
             Action<IEndpointConfigurationV2> destinationConfig = c =>
             {
                 c.UseConnectionString(ConnectionStrings.Default);
             };
 
-            VerifySend("3.0", sourceConfig, "2.2", destinationConfig);
+            VerifyRoundtrip("3.0", sourceConfig, "2.2", destinationConfig);
         }
 
-        void VerifySend<S, D>(string sourceVersion, Action<S> sourceConfig, string destinationVersion, Action<D> destinationConfig)
+        void VerifyRoundtrip<S, D>(string initiatorVersion, Action<S> initiatorConfig, string replierVersion, Action<D> replierConfig)
             where S : IEndpointConfiguration
             where D : IEndpointConfiguration
         {
-            using (var source = EndpointFacadeBuilder.CreateAndConfigure(sourceEndpointDefinition, sourceVersion, sourceConfig))
-            {
-                using (var destination = EndpointFacadeBuilder.CreateAndConfigure(destinationEndpointDefinition, destinationVersion, destinationConfig))
-                {
-                    var messageId = Guid.NewGuid();
 
-                    source.SendCommand(messageId);
+            using (var source = EndpointFacadeBuilder.CreateAndConfigure(sourceEndpointDefinition, initiatorVersion, initiatorConfig))
+            {
+                using (EndpointFacadeBuilder.CreateAndConfigure(destinationEndpointDefinition, replierVersion, replierConfig))
+                {
+                    var requestId = Guid.NewGuid();
+
+                    source.SendRequest(requestId);
 
                     // ReSharper disable once AccessToDisposedClosure
-                    AssertEx.WaitUntilIsTrue(() => destination.ReceivedMessageIds.Any(mi => mi == messageId));
+                    AssertEx.WaitUntilIsTrue(() => source.ReceivedResponseIds.Any(responseId => responseId == requestId));
                 }
             }
         }
